add call to process a cards login in process turn
this actually turned out to be a HUGE pain. The logic that verifies that a card has all the parameters it needs to be processed broke a ton of tests. With good reason, because the players in teh tests were jsut discarding whatever they had. I changed the seeds for those tests so that the card is always a countess, because it doesn't have any discard logic so it shouldn't affect the existing tests at all