Notify each player when it is their turn
This was a surprisingly large refactor. The return type of process turn
was changed to that winners were included as messages rather than
separate entries in the tuple.

Right now the sequence of messages is combined in a way that makes all
messages are public. That will need to be fixed. Fixing that will
involve separating out the response url from the slack request so that
multiple responses can be processed.

this also broke like all the unit tests. Will fix soon I promise