refactor process turn with case classes
to remove the actual messages from the game engine and make it so we don't have to compare strings meant for humans in tests, have the results of processTurn be one of a few cases instead of a list of messages

for some reason, some uses of `inside` were causing weird jvm runtime errors, saying the method had already been loaded. I had to replace 2 of them to not use `inside` and now it works.