add jester behavior cards
refactor checkMatchOver to return a list of players who got points

cards like jester mean that potentially more than one player could receive points at the end of a match. I've decided to go with the convention of putting the match winner first in the list, and any other players who received points come after. We pull the list apart into a match winner and other points when processing the turn and converting to a turn result.

going to have to do this to GameWinner too, since potentially we could have multiple winners now